requirements.txt
django
django_extensions
django-registration
django-debug-toolbar
ipython>=0.10
ipdb
pep8==0.6.1
django-tastypie
Pillow
django-haystack
